LPD-33314 Add new endpoints
