--- modules/dxp/apps/analytics/analytics-reports-js-components-web/src/main/resources/META-INF/resources/js/apis/utils.ts
@@ -1,6 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-export const API_PREFIX = '/o/analytics-reports-rest/v1.0';
